# openapi/v1/api.yaml
openapi: 3.0.0
paths:
...
  /v1/permissions:
    get:
      operationId: PermissionsController_findAll
      summary: Listar todas as permissões disponíveis no sistema
      parameters: []
      responses:
        '200':
          description: Lista de permissões do sistema
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
      tags:
        - roles
      security:
        - bearer: []
...
  /v1/occurrences:
    get:
      operationId: OccurrencesController_findAll
      summary: List all occurrences with cursor-based pagination
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - occurrences
      security:
        - bearer: []
    post:
      operationId: OccurrencesController_create
      summary: Create a new occurrence
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - occurrences
      security:
        - bearer: []
  /v1/occurrences/{id}:
    get:
      operationId: OccurrencesController_findById
      summary: Get occurrence by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - occurrences
      security:
        - bearer: []
    patch:
      operationId: OccurrencesController_update
      summary: Update occurrence fields (not status)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - occurrences
      security:
        - bearer: []
  /v1/occurrences/{id}/status:
    patch:
      operationId: OccurrencesController_changeStatus
      summary: Change occurrence status
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - occurrences
      security:
        - bearer: []
  /v1/occurrences/{id}/resolve:
    patch:
      operationId: OccurrencesController_resolve
      summary: Resolve an occurrence with action taken
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - occurrences
      security:
        - bearer: []
  /v1/occurrences/{id}/escalate:
    patch:
      operationId: OccurrencesController_escalate
      summary: Escalate occurrence to CRITICAL
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - occurrences
      security:
        - bearer: []
...
  /v1/maintenance:
    get:
      operationId: MaintenanceController_findAll
      summary: List maintenance orders with cursor pagination
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
    post:
      operationId: MaintenanceController_create
      summary: Create a new maintenance order
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
  /v1/maintenance/preventive-due:
    get:
      operationId: MaintenanceController_getPreventiveDue
      summary: List vehicles with overdue preventive maintenance (> 90 days)
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
  /v1/maintenance/{id}:
    get:
      operationId: MaintenanceController_findById
      summary: Get maintenance order by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
    patch:
      operationId: MaintenanceController_update
      summary: Update maintenance order fields
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
    delete:
      operationId: MaintenanceController_cancel
      summary: Cancel maintenance order
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
  /v1/maintenance/{id}/complete:
    patch:
      operationId: MaintenanceController_complete
      summary: Mark maintenance order as completed
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
  /v1/maintenance/{id}/items:
    post:
      operationId: MaintenanceController_addItem
      summary: Add item to maintenance order
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
  /v1/maintenance/{id}/items/{itemId}:
    delete:
      operationId: MaintenanceController_removeItem
      summary: Remove item from maintenance order
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: itemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - maintenance
      security:
        - bearer: []
...
  /v1/notifications:
    get:
      operationId: NotificationsController_findAll
      summary: List notifications for the authenticated user
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - notifications
      security:
        - bearer: []
  /v1/notifications/unread-count:
    get:
      operationId: NotificationsController_findUnreadCount
      summary: Get count of unread notifications for the authenticated user
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - notifications
      security:
        - bearer: []
  /v1/notifications/read-all:
    patch:
      operationId: NotificationsController_markAllAsRead
      summary: Mark all notifications as read for the authenticated user
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - notifications
      security:
        - bearer: []
  /v1/notifications/{id}/read:
    patch:
      operationId: NotificationsController_markAsRead
      summary: Mark a single notification as read
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - notifications
      security:
        - bearer: []
...
info:
  title: Gestão Fretamento Pro API
  description: API REST para gestão de fretamento empresarial
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...